Remove "Appearance" from Custom View.
This allows building on 10.8.

Note that opening the plugin window causes Xcode to crash, however I
suspect the crash is unrelated as removing this attribute is the same as
not specifying a custom 'Appearance' in the first place.

Related, IB in Xcode 5.0.2 on 10.8 has no 'Appearance' option for
NSViews, where as Xcode 5.0.2 on 10.9 does.