OSX 10.10 compatibility fixes.
The app wasn't able to download its package list.
The request was seemingly correct, but the response contained 0 bytes.
Since GitHub specifically forbids using raw.github.com as a CDN, perhaps
that is the reason why the request failed.
Since the package list was a GitHub repository, and the app contained
code for cloning repositories, the solution is trivial.
The package repository is now being cloned/updated on startup, and the
packages list is taken from there instead of being downloaded directly.

Additionally, the app was crashing when accessing
"[[NSOperationQueue mainQueue] operations].lastObject".
It was trying to add a dependency to an NSOperation, but it was nil.
The dependency was removed and a package that is updating will no longer
wait for the previous package to finish updating.